Added filtering by States and Colors, removed ushort because it caused an error in query, maded filter spec clearer

+++ src/PetAdoptionApp.Api/Models/PostPetProfileRequest.cs
@@ -12,4 +12,4 @@ public record PostPetProfileRequest(
 	int SizeId,
 	List<int>? ColorIds,
 	PetProfileDetailsCommandDto Details,
-	List<ushort>? States);
+	List<int>? States);
