+++ src/PetAdoptionApp.Domain/Aggregates/StateAggregate/Enums/Status.cs
@@ -2,7 +2,7 @@
 
 namespace PetAdoptionApp.Domain.Aggregates.StateAggregate.Enums;
 
-public class Status : SmartEnum<Status, ushort>
+public class Status : SmartEnum<Status, int>
 {
 	public static readonly Status NeedsHome = new ("Needs home", 1);
 	public static readonly Status NeedsSitter = new("Needs sitter", 2);
@@ -11,7 +11,7 @@ public class Status : SmartEnum<Status, ushort>
 
 	public static readonly Status NeedsDonation = new("Needs donation", 5);
 
-	private Status(string name, ushort value) : base(name, value) { }
+	private Status(string name, int value) : base(name, value) { }
 
 
 	private static readonly IReadOnlyCollection<Status> _conflictedStatuses =
@@ -20,6 +20,6 @@ public class Status : SmartEnum<Status, ushort>
 	public static List<Status> FilterConflicted(List<Status> statusList) =>
 		statusList.Intersect(_conflictedStatuses).ToList();
 
-	public static List<ushort> FilterConflicted(List<ushort> statusList) =>
+	public static List<int> FilterConflicted(List<int> statusList) =>
 		statusList.Intersect(_conflictedStatuses.Select(s => s.Value)).ToList();
 }
